add post estimation diagnostic test
Add diagnostics for consideration of the veracity of the assumptions, especially to confirm that the error terms are distributed as assumed. `gam.check(fit1)` performs several standard regression diagnostic tests and is in addition to the other model selection procedures that were already described here such as data visualisation and the comparison of Akaike Information Criteria (AIC) across a set of candidate models.